better slice handling and test cases moved
